project-3-implementation-sdes/4chan/requirements.txt
###### Requirements with Version Specifiers ######
pyfaktory
python-dotenv ~= 1.0
pymongo
requests ~= 2.32
pandas
matplotlib
